requirements.txt
python-telegram-bot>=21.10
openai>=1.99.0
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.37.0
openpyxl==3.1.5
fastapi>=0.115.0
uvicorn>=0.30.0
